Extract the business logic of the version command into a service
